# expression: implement vectorized evaluation for `builtinSubDateAndStringSig`