@wshwsh12 2019-12-04
Thanks for your contribution. Sorry to reply you so late. 
We can use go generate to implement the following functions at the same time. Would you like to try it?
You can see the docs https://docs.google.com/document/d/17XdYo_EwYaWANqbjS3i3OVSz3oLlovU2fIo1bRx4YqA/edit# to get more information.
```
builtinSubDateAndDurationSig
builtinSubDateAndStringSig
builtinSubDatetimeAndDurationSig
builtinSubDatetimeAndStringSig
builtinSubDurationAndDurationSig
builtinSubDurationAndStringSig
builtinSubStringAndDurationSig
builtinSubStringAndStringSig
builtinSubTimeDateTimeNullSig
builtinSubTimeDurationNullSig
builtinSubTimeStringNullSig
```