@wshwsh12 2019-12-04
> I will try it.

@shihongzhi  Sorry, but .... someone is doing it . The pr is https://github.com/pingcap/tidb/pull/13691.